fix: QuestionAndAnswer가 DTO에 의존하는 이슈 해결
